Move `HostNameUtils` to :core:common
