+++ feature/autodiscovery/autoconfig/build.gradle.kts
@@ -7,6 +7,7 @@ dependencies {
     api(projects.feature.autodiscovery.api)
 
     compileOnly(libs.xmlpull)
+    implementation(projects.core.common)
     implementation(libs.okhttp)
     implementation(libs.minidns.hla)
